fix(db): split chunk_stats test into sibling file (size cap)
tests.rs hit 1008 lines after the pending_chunks regression test was
added. Move the new test to chunk_stats_tests.rs (sibling pattern
already used by audit_tests, drift_tests, etc.) so tests.rs drops to
979 lines and the file-size CI gate passes again.

Bump version to 0.3.85.
