+++ crates/rs-core/src/db/mod.rs
@@ -64,6 +64,9 @@ mod streaming_event_flag_tests;
 #[cfg(test)]
 mod v2_tests;
 
+#[cfg(test)]
+mod chunk_stats_tests;
+
 /// Create a SQLite connection pool.
 pub async fn create_pool(db_path: &Path) -> Result<SqlitePool> {
     let url = format!("sqlite:{}?mode=rwc", db_path.display());
